// continental-frontend/src/interfaces/Api.interface.ts
export interface ApiResponse<T = unknown> {
  data?: T;
  message?: string;
  success: boolean;
  errorMsg?: string;
}